requirements.txt
pandas
numpy
gspread
google-auth
sqlalchemy
psycopg2-binary
python-dotenv
